backend/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# Validation & schemas
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database & Auth
supabase>=2.3.0

# AI / LLM (Google AI Studio / Gemini)
langchain>=0.1.0
langchain-google-genai>=2.0.0

# Payments
stripe>=8.0.0

# Email (alertas)
resend>=2.0.0

# Document generation
reportlab>=4.0.0        # PDF puro Python (reemplaza weasyprint que necesita GTK3 en Windows)
python-docx>=1.1.0
PyPDF2>=3.0.0          # Para leer PDFs
pdfplumber>=0.9.0      # Alternativa más robusta para PDFs
Pillow>=10.0.0         # Para procesar imágenes
pytesseract>=0.3.10    # OCR para extraer texto de imágenes
pdf2image>=1.16.0      # Convertir PDF a imágenes para OCR

# Auth
PyJWT>=2.8.0

# Testing
pytest>=8.0.0

# Utilities
python-multipart>=0.0.6
sse-starlette>=1.8.0
httpx>=0.26.0
structlog>=23.2.0
